feat: modal 컴포넌트 리팩토링 (#492)
* feat: Portal 컴포넌트 제작

* refactor: Modal 컴포넌트 Content와 합친 후 변경사항 대응 (#490)

* Squashed commit of the following:

commit 32a34b6a65e72630c5ec5f1973bc1ff1acb7841f
Author: 황준승 <78203399+turtle601@users.noreply.github.com>
Date:   Wed Sep 20 14:40:45 2023 +0900

    feat: 이미지 최적화 (#485)

    * feat: 사용하지 않는 .png 파일 제거 (#478)

    * refactor: Webpack 파일 리팩토링 (#478)

    * refactor: 불필요한 코드 구문 제거 (#478)

    타입스크립트 컴파일은 ts loader로만으로도 충분

    * refactor: devDependency vs Dependency 라이브러리 분리 (#478)

    * refactor: 불필요한 코드 삭제 (#478)

    Changed:
    불필요한 optimize 구문 삭제'

    * feat: picture 태그를 활용한 이미지 최적화 (#478)

    * refactor: 불필요한 react.memo 제거 (#478)

    * feat: 이미지 url 관련 유틸함수 구현 (#478)

    * feat: 폰트 리로드 적용 (#478)

    * feat: video lazy로딩 적용 (#478)

    * fix: restaurants mock 데이터 수정(#478)
    Changed:
    .jpeg 삭제

    * refactor: 불필요한 폰트 리로딩, lazy loading 제거 (#478)

    * fix: react-query-devtools 디펜던시 오류 해결 (#478)

    * fix: 빌드 시 lint 오류 해결 (#478)

    Changed:
    eslint 설정에 dependency 설정 끔

commit 276eec9f9fa225f08639947a037b8d2b5daebd74
Author: Minjae Kim <smallkdb@gmail.com>
Date:   Wed Sep 20 14:38:43 2023 +0900

    feat: loading 상태에 필요한 skeleton 제작 (#488)

    * feat: 데이터 추가 mock api 연결

    * refactor: 인증 필요 없는 기능 client 변경

    * feat: mse 동영상, 리뷰 기능 구현 (#472)

    * refactor: api 훅 함수로 분리, 적용 (#472)

    * refactor: 쓰지 않는 파일 삭제 (#472)

    * fix: console log 삭제 (#472)

    * Squashed commit of the following:

    commit 6e0068735b155b95ec45de7fd78eb241993dc513
    Author: Jeremy <102432453+shackstack@users.noreply.github.com>
    Date:   Fri Sep 15 17:18:33 2023 +0900

        bug: 옵션관련 ux 개선 (#474)

        * fix: 검색창 옵션 수정 (#473)

        * fix: 지도 줌 limit 설정 (#473)

        * fix: 모바일 환경에서 거리순으로 설정 (#473)

        * design: 오버레이 마커 디자인 개선 (#473)

    * fix: 오류 수정 (#476)

    * refactor: useScrollBlock 보강 (#476)

    * refactor:  모바일 페이지 컴포넌트 분리 (#476)

    * refactor:  라우터 재설정 (#476)

    * fix: 잘못된 import 수정 (#476)

    * refactor: router 및 outlet 새로 설정 (#476)

    * feat: NavBar 관련 Skeleton 제작 (#480)

    * feat: NavBar 관련 Skeleton 제작 (#480)

    * refactor: navBarSkeleton의 NavItem 개수를 props로 받도록 설정 (#480)

* refactor: Login 관련 Modal 자체 모달로 분리 (#490)

* refactor: 리뷰 관련 모달 로직 수정 (#490)

* Squashed commit of the following:

commit ac40d18c9959fb488c8bded376ff554429bbf712
Author: 황준승 <78203399+turtle601@users.noreply.github.com>
Date:   Wed Sep 20 16:01:35 2023 +0900

    [✨feat] 폰트 + 이미지 최적화 (#491)

    * feat: 사용하지 않는 .png 파일 제거 (#478)

    * refactor: Webpack 파일 리팩토링 (#478)

    * refactor: 불필요한 코드 구문 제거 (#478)

    타입스크립트 컴파일은 ts loader로만으로도 충분

    * refactor: devDependency vs Dependency 라이브러리 분리 (#478)

    * refactor: 불필요한 코드 삭제 (#478)

    Changed:
    불필요한 optimize 구문 삭제'

    * feat: picture 태그를 활용한 이미지 최적화 (#478)

    * refactor: 불필요한 react.memo 제거 (#478)

    * feat: 이미지 url 관련 유틸함수 구현 (#478)

    * feat: 폰트 리로드 적용 (#478)

    * feat: video lazy로딩 적용 (#478)

    * fix: restaurants mock 데이터 수정(#478)
    Changed:
    .jpeg 삭제

    * refactor: 불필요한 폰트 리로딩, lazy loading 제거 (#478)

    * feat: 폰트 preload 최적화 (#487)

    * refactor: devDependency 허용하는 eslint 설정 추가 (#487)

    * fix: webp이미지 불러오지 못하는 오류 해결 (#487)

    * refactor: 불필요한 코드 삭제 (#478)

    Changed:
    ${} 삭제

    * refactor: url 전역 변수 선언 (#487)

    * refactor: React.MouseEvent 코드 컨벤션 수정 (#487)

    * fix: 이미지 .. 붙여지는 오류 해결 (#487)

    * fix: react-query devtools 디펜던시 lint 에러 해결 (#487)

commit 32a34b6a65e72630c5ec5f1973bc1ff1acb7841f
Author: 황준승 <78203399+turtle601@users.noreply.github.com>
Date:   Wed Sep 20 14:40:45 2023 +0900

    feat: 이미지 최적화 (#485)

    * feat: 사용하지 않는 .png 파일 제거 (#478)

    * refactor: Webpack 파일 리팩토링 (#478)

    * refactor: 불필요한 코드 구문 제거 (#478)

    타입스크립트 컴파일은 ts loader로만으로도 충분

    * refactor: devDependency vs Dependency 라이브러리 분리 (#478)

    * refactor: 불필요한 코드 삭제 (#478)

    Changed:
    불필요한 optimize 구문 삭제'

    * feat: picture 태그를 활용한 이미지 최적화 (#478)

    * refactor: 불필요한 react.memo 제거 (#478)

    * feat: 이미지 url 관련 유틸함수 구현 (#478)

    * feat: 폰트 리로드 적용 (#478)

    * feat: video lazy로딩 적용 (#478)

    * fix: restaurants mock 데이터 수정(#478)
    Changed:
    .jpeg 삭제

    * refactor: 불필요한 폰트 리로딩, lazy loading 제거 (#478)

    * fix: react-query-devtools 디펜던시 오류 해결 (#478)

    * fix: 빌드 시 lint 오류 해결 (#478)

    Changed:
    eslint 설정에 dependency 설정 끔

commit 276eec9f9fa225f08639947a037b8d2b5daebd74
Author: Minjae Kim <smallkdb@gmail.com>
Date:   Wed Sep 20 14:38:43 2023 +0900

    feat: loading 상태에 필요한 skeleton 제작 (#488)

    * feat: 데이터 추가 mock api 연결

    * refactor: 인증 필요 없는 기능 client 변경

    * feat: mse 동영상, 리뷰 기능 구현 (#472)

    * refactor: api 훅 함수로 분리, 적용 (#472)

    * refactor: 쓰지 않는 파일 삭제 (#472)

    * fix: console log 삭제 (#472)

    * Squashed commit of the following:

    commit 6e0068735b155b95ec45de7fd78eb241993dc513
    Author: Jeremy <102432453+shackstack@users.noreply.github.com>
    Date:   Fri Sep 15 17:18:33 2023 +0900

        bug: 옵션관련 ux 개선 (#474)

        * fix: 검색창 옵션 수정 (#473)

        * fix: 지도 줌 limit 설정 (#473)

        * fix: 모바일 환경에서 거리순으로 설정 (#473)

        * design: 오버레이 마커 디자인 개선 (#473)

    * fix: 오류 수정 (#476)

    * refactor: useScrollBlock 보강 (#476)

    * refactor:  모바일 페이지 컴포넌트 분리 (#476)

    * refactor:  라우터 재설정 (#476)

    * fix: 잘못된 import 수정 (#476)

    * refactor: router 및 outlet 새로 설정 (#476)

    * feat: NavBar 관련 Skeleton 제작 (#480)

    * feat: NavBar 관련 Skeleton 제작 (#480)

    * refactor: navBarSkeleton의 NavItem 개수를 props로 받도록 설정 (#480)

* chore: lock 파일 변경 (#490)

* fix: lint 에러 수정 (#490)
