+++ frontend/src/components/@common/Modal/index.tsx
@@ -1,4 +1,3 @@
 import Modal from '~/components/@common/Modal/Modal';
-import ModalContent from '~/components/@common/Modal/ModalContent';
 
-export { Modal, ModalContent };
+export default Modal;
