+++ frontend/src/components/@common/Portal/index.tsx
@@ -1,0 +1,3 @@
+import Portal from './Portal';
+
+export default Portal;
