+++ frontend/src/components/LoginModal/index.tsx
@@ -1,0 +1,3 @@
+import LoginModal from '~/components/LoginModal/LoginModal';
+
+export default LoginModal;
