--- frontend/src/components/LoginModalContent/index.tsx
@@ -1,3 +1,0 @@
-import LoginModalContent from '~/components/LoginModalContent/LoginModalContent';
-
-export default LoginModalContent;
