+++ frontend/src/components/RestaurantCard/RestaurantCard.tsx
@@ -8,8 +8,7 @@ import { FONT_SIZE, truncateText } from '~/styles/common';
 
 import type { Celeb } from '~/@types/celeb.types';
 import type { Restaurant } from '~/@types/restaurant.types';
-import { Modal, ModalContent } from '~/components/@common/Modal';
-import LoginModalContent from '~/components/LoginModalContent';
+import LoginModal from '~/components/LoginModal';
 import useToggleLikeNotUpdate from '~/hooks/server/useToggleLikeNotUpdate';
 
 interface RestaurantCardProps {
@@ -22,7 +21,7 @@ interface RestaurantCardProps {
 
 function RestaurantCard({ restaurant, celebs, size, type = 'list', setHoveredId = () => {} }: RestaurantCardProps) {
   const { id, images, name, roadAddress, category, phoneNumber } = restaurant;
-  const { isModalOpen, closeModal, openModal, toggleRestaurantLike, isLiked } = useToggleLikeNotUpdate(restaurant);
+  const { isModalOpen, closeModal, toggleRestaurantLike, isLiked } = useToggleLikeNotUpdate(restaurant);
 
   const navigate = useNavigate();
 
@@ -66,11 +65,8 @@ function RestaurantCard({ restaurant, celebs, size, type = 'list', setHoveredId 
           </StyledProfileImageSection>
         </StyledInfoSection>
       </StyledContainer>
-      <Modal isOpen={isModalOpen} close={closeModal} open={openModal}>
-        <ModalContent title="로그인 및 회원 가입">
-          <LoginModalContent />
-        </ModalContent>
-      </Modal>
+
+      <LoginModal isOpen={isModalOpen} close={closeModal} />
     </>
   );
 }
