+++ frontend/src/pages/SignUpPage.tsx
@@ -1,12 +1,23 @@
-import LoginModalContent from '~/components/LoginModalContent';
+import styled from 'styled-components';
+import LoginButton from '~/components/@common/LoginButton';
 import LoginPageUI from '~/components/LoginPageUI';
 
 function SignUpPage() {
   return (
     <LoginPageUI title="로그인 및 회원가입">
-      <LoginModalContent />
+      <StyledLoginContainer>
+        <LoginButton type="kakao" />
+        <LoginButton type="google" />
+      </StyledLoginContainer>
     </LoginPageUI>
   );
 }
 
 export default SignUpPage;
+
+const StyledLoginContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+
+  margin-bottom: 1.6rem;
+`;
